+++ database/migrations/2022_04_17_133712_remove_tahapan_from_blok_lahan.php
@@ -13,10 +13,13 @@ class RemoveTahapanFromBlokLahan extends Migration
      */
     public function up()
     {
-        Schema::table('blok_lahan', function (Blueprint $table) {
+        $driver = Schema::connection($this->getConnection())->getConnection()->getDriverName();
+        Schema::table('blok_lahan', function (Blueprint $table) use ($driver) {
             //
-            $table->dropForeign(['tahapan_id']);
-            $table->dropColumn('tahapan_id');
+            if($driver !== 'sqlite'){
+                $table->dropForeign(['tahapan_id']);
+                $table->dropColumn('tahapan_id');
+            }
         });
     }
